refactor: BankManagerConsole 파일 분리, 접근제어자 설정 #4
